Change default sort hook:
- Make assoc sort hook work with default sort params
- Move simple sort to custom hooks
- Modify tests for the above changes
