+++ test/custom_hooks/simple_sort_test.exs
@@ -1,0 +1,5 @@
+defmodule Rummage.Ecto.CustomHooks.SimpleSortTest do
+  use ExUnit.Case
+  doctest Rummage.Ecto.CustomHooks.SimpleSort
+
+end
