+++ test/hooks/paginate_test.exs
@@ -1,4 +1,4 @@
-defmodule Rummage.Ecto.PaginateTest do
+defmodule Rummage.Ecto.Hooks.PaginateTest do
   use ExUnit.Case
   doctest Rummage.Ecto.Hooks.Paginate
