+++ test/hooks/sort_test.exs
@@ -1,4 +1,4 @@
-defmodule Rummage.Ecto.SortTest do
+defmodule Rummage.Ecto.Hooks.SortTest do
   use ExUnit.Case
   doctest Rummage.Ecto.Hooks.Sort
